Fix Hybrid Engine metrics printing (#4789)
The metrics print statement in the Hybrid Engine had a bug where
`self._total_batch_size` was assumed to exist even when this variable
was not initialized due to no corresponding inference policy.

This PR fixes this by performing a check when constructing the metrics
print.

Addresses: https://github.com/microsoft/DeepSpeedExamples/issues/593