Monitor was always enabled causing performance degradation (#5633)
The Boolean expression for the monitor to be enabled was incorrect, as
instead of using the `enabled` field, it used the comet configuration
object, making the expression always True.

This caused performance degradation (we've observed ~10% drop) as it
erroneously invoked the events logging flow along with the expensive
calculation of `loss.mean().item()`.

Co-authored-by: Olatunji Ruwase <olruwase@microsoft.com>
Co-authored-by: Logan Adams <114770087+loadams@users.noreply.github.com>